pyproject.toml
[project]
name = "kuroko"
version = "0.1.0"
description = "A CLI tool to collect and visualize LLM agent checkpoints (kuroko)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "PyYAML>=6.0",
    "pydantic>=2.0",
    "pytest>=7.0",
]

[project.scripts]
kuroko = "kuroko.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
